created teleop class and moveed around code
